+++ .changeset/a2d40a44/changes.md
@@ -1,1 +1,1 @@
-- Added pagination to package changelogs. The default number of logs per page is 20. If null is given to maxItems prop then the changelog is not paginated.
+- Added pagination to package changelogs. The default number of logs per page is 20. If null is given to entriesPerPage prop then the changelog is not paginated.